requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.5
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.6
catboost==1.2.7
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
cmake==4.2.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
contourpy==1.3.1
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.2.0
deepface==0.0.99
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.8
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib-bin==20.0.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.2.0
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.62.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphviz==0.20.3
greenlet==3.3.1
groq==0.37.1
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==25.1.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.10
imbalanced-learn==0.14.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.13.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.5
keras==3.10.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-classic==1.0.1
langchain-community==0.3.31
langchain-core==0.3.83
langchain-google-genai==2.0.10
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightdsa==0.0.3
lightecc==0.0.4
lightgbm==4.6.0
lightphe==0.0.20
llvmlite==0.44.0
lz4==4.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.24.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.0
openai==1.109.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.0.0
pipenv==2024.1.0
platformdirs==4.3.6
plotly==6.0.0
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pyparsing==3.2.1
pypdf==6.7.0
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-slugify==8.0.4
pyttsx3==2.98
pytz==2024.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.3
pyzmq==26.2.1
razorpay==1.4.2
referencing==0.36.2
regex==2025.9.1
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.2.2
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.46
sqlparse==0.5.2
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.20.1
threadpoolctl==3.5.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.2
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.1.0
typer==0.22.0
typer-slim==0.22.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
utils==1.0.2
uuid_utils==0.14.0
validators==0.34.0
virtualenv==20.26.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==2.1.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
